splitting out scal tests into folder
