+++ app/tests/scal/test_scal_med.py
@@ -1,0 +1,76 @@
+"""
+Script to run through consumer tests for GP CONNECT Scal
+"""
+
+import json
+
+import pytest
+
+from app.gpconnect import gpconnect
+
+from ..log_context import capture_test_logs
+
+audit_dict = {
+    "subject_id": "CONE, Stephen",
+    "organization": "UCLH - University College London Hospitals - TST",
+    "organization_id": "urn:oid:1.2.840.114350.1.13.525.3.7.3.688884.100",
+    "home_community_id": "urn:oid:1.2.840.114350.1.13.525.3.7.3.688884.100",
+    "role": {
+        "Role": {
+            "@codeSystem": "2.16.840.1.113883.6.96",
+            "@code": "224608005",
+            "@codeSystemName": "SNOMED_CT",
+            "@displayName": "Administrative healthcare staff",
+            "@xmlns": "urn:hl7-org:v3",
+            "@xmlns:xsi": "http://www.w3.org/2001/XMLSchema-instance",
+            "@xmlns:xsd": "http://www.w3.org/2001/XMLSchema",
+        }
+    },
+    "purpose_of_use": {
+        "PurposeForUse": {
+            "@xsi:type": "CE",
+            "@code": "TREATMENT",
+            "@codeSystem": "2.16.840.1.113883.3.18.7.1",
+            "@codeSystemName": "nhin-purpose",
+            "@displayName": "Treatment",
+            "@xmlns": "urn:hl7-org:v3",
+            "@xmlns:xsi": "http://www.w3.org/2001/XMLSchema-instance",
+            "@xmlns:xsd": "http://www.w3.org/2001/XMLSchema",
+        },
+    },
+    "resource_id": "9690937278^^^&2.16.840.1.113883.2.1.4.1&ISO",
+}
+
+
+@pytest.mark.asyncio
+async def test_GPC_STR_TST_MED_02():
+    """Given I have received a successful, valid medications message response
+    When I display or use the medication information
+    Then I display or utilise all the key information to represent or process the medication record(s) commenserate with the original record meaning and my specific use case
+    """
+    nhsnos = ["9690937286"]
+    for nhsno in nhsnos:
+        async with capture_test_logs("GPC-STR-TST-MED-02", nhsno) as log_dir:
+            result = await gpconnect(nhsno, saml_attrs=audit_dict, log_dir=log_dir)
+            code = result.status_code
+            body = json.loads(result.body)
+            assert "document_id" in body
+
+
+@pytest.mark.asyncio
+async def test_GPC_STR_TST_MED_07():
+    """
+    Given I have received a successful, valid medications message response
+    And the response has a list with an empty reason
+    And the response does not include medication resourcese
+    When I display or use the medication information
+    Then I display or utilise the list empty reson to inform the user that the patient has no medication records within the request parameters in a way appropriate to my use case
+    """
+    nhsnos = ["9690937308"]
+    for nhsno in nhsnos:
+        async with capture_test_logs("GPC-STR-TST-MED-07", nhsno) as log_dir:
+            result = await gpconnect(nhsno, saml_attrs=audit_dict, log_dir=log_dir)
+            code = result.status_code
+            body = json.loads(result.body)
+            assert "document_id" in body
+
