+++ app/tests/scal/test_scal_spine.py
@@ -10,8 +10,8 @@ import pytest
 from app.gpconnect import gpconnect
 from app.pds.pds import lookup_patient, sds_trace
 
-from .log_context import capture_test_logs
-from .test_scal import audit_dict
+from app.tests.log_context import capture_test_logs
+from .test_scal_alg import audit_dict
 
 
 @pytest.mark.asyncio
