+++ pytest.ini
@@ -1,4 +1,3 @@
 [pytest]
 addopts = 
-    --ignore=app/tests/test_scal.py
-    --ignore=app/tests/test_scal_spine.py
+    --ignore=app/tests/scal/
